mudando a tipagem do objeto

+++ kilowatts/src/Kilowatts.java
@@ -4,9 +4,9 @@ public class Kilowatts {
     public static void main(String[] args) {
         Scanner teclado = new Scanner(System.in);
         System.out.println("Informe a quantidade de kilowatts consumidos na casa: ");
-        int qtdKilowatt = teclado.nextInt();
+        double qtdKilowatt = teclado.nextDouble();
         System.out.println("Informe o valor do kilowatt: ");
-        int valKilowatt = teclado.nextInt();
+        double valKilowatt = teclado.nextDouble();
 
         if (qtdKilowatt < 150) {
             System.out.println("Valor a ser pago: " + (qtdKilowatt * valKilowatt * 0.90));
